fix: waitWorkflowCompletion operation will return a DTO instead of just the enum
